Added test, some Provider cleanup for order page.
