+++ src/Tickets/Commerce/Hooks.php
@@ -24,7 +24,6 @@ use TEC\Tickets\Commerce\Admin_Tables\Orders_Table;
 use TEC\Tickets\Commerce\Reports\Orders;
 use TEC\Tickets\Commerce\Status\Completed;
 use TEC\Tickets\Commerce\Status\Status_Interface;
-use TEC\Tickets\Commerce\Payments_Tab;
 use TEC\Tickets\Commerce\Status\Status_Handler;
 use WP_Admin_Bar;
 use Tribe__Date_Utils;
@@ -101,8 +100,6 @@ class Hooks extends Service_Provider {
 		add_action( 'pre_get_posts', [ $this, 'pre_filter_admin_order_table' ] );
 
 		add_action( 'admin_menu', tribe_callback( Orders_Page::class, 'add_orders_page' ), 15 );
-
-		add_action( 'current_screen', [ $this, 'breadcrumb_order_edit_screen' ] );
 	}
 
 	/**
@@ -139,43 +136,13 @@ class Hooks extends Service_Provider {
 		$this->provider_meta_sanitization_filters();
 
 		add_filter( 'tribe_template_context:tickets-plus/v2/tickets/submit/button-modal', [ $this, 'filter_showing_cart_button' ] );
-
 		add_filter( 'tec_tickets_commerce_payments_tab_settings', [ $this, 'filter_payments_tab_settings' ] );
-
 		add_filter( 'wp_redirect', [ $this, 'filter_redirect_url' ] );
-
 		add_filter( 'tec_tickets_editor_configuration_localized_data', [ $this, 'filter_block_editor_localized_data' ] );
-
 		add_action( 'tribe_editor_config', [ $this, 'filter_tickets_editor_config' ] );
-
 		add_filter( 'wp_list_table_class_name', [ $this, 'filter_wp_list_table_class_name' ], 10, 2 );
-
 		add_filter( 'tribe_dropdown_tec_tc_order_table_events', [ $this, 'provide_events_results_to_ajax' ], 10, 2 );
-
 		add_filter( 'tribe_dropdown_tec_tc_order_table_customers', [ $this, 'provide_customers_results_to_ajax' ], 10, 2 );
-	}
-
-	public function breadcrumb_order_edit_screen() {
-		$screen = get_current_screen();
-		if ( ! $screen
-			|| $screen->id !== Order::POSTTYPE
-			|| $screen->post_type !== Order::POSTTYPE
-			|| $screen->base !== 'post' ) {
-			return;
-		}
-
-		add_action( 'all_admin_notices', [ $this, 'print_breadcrumb_order_edit_screen_html' ], 999 );
-	}
-
-	public function print_breadcrumb_order_edit_screen_html() {
-		$url  = get_current_screen()->parent_file;
-		$text = _x( 'Orders', 'Order edit page breadcrumb link back to Orders page.', 'event-tickets' );
-
-		echo <<<STR
-		<div class="tec-tickets-commerce-single-order--breadcrumb--order--edit">
-			<a href="$url"><span class="dashicons dashicons-arrow-left-alt2"></span> $text</a>
-		</div>
-STR;
 	}
 
 	/**
