+++ src/resources/postcss/tickets-commerce/admin/orders/single.pcss
@@ -287,12 +287,17 @@ body.wp-admin.post-type-tec_tc_order {
 			}
 		}
 	}
-	&--breadcrumb--order--edit a {
-		text-decoration: none;
-		.dashicons {
-			font-size: var(--tec-font-size-2);
-			line-height: var(--tec-font-size-3);
-			vertical-align: middle;
+	&--breadcrumb--order--edit {
+		margin-top: 1em;
+
+		a {
+			text-decoration: none;
+
+			.dashicons {
+				font-size: var(--tec-font-size-2);
+				line-height: var(--tec-font-size-4);
+				vertical-align: middle;
+			}
 		}
 	}
 }
